@@ -1,22 +1,17 @@
-const {
-  describe,
-  it
-} = require('mocha')
-const { expect } = require('chai')
-const { getChampionName, getChampionNameCapital } = require('../../src//constants/champions')
+import { getChampionName, getChampionNameCapital } from '../../src//constants/champions'
 
 describe('Champions Constant', () => {
   it('should return a empty champ', () => {
     const response = getChampionName(0)
-    expect(response).to.eq('EMPTY_CHAMPION')
+    expect(response).toEqual('EMPTY_CHAMPION')
   })
   it('should return a champ', () => {
     const response = getChampionName(1)
-    expect(typeof response).to.eq('string')
+    expect(typeof response).toEqual('string')
   })
   it('should return a champ capitalize', () => {
     const response = getChampionNameCapital(4)
-    expect(response).to.eq('TwistedFate')
+    expect(response).toEqual('TwistedFate')
   })
   it('should return error unknow champion', (done) => {
     try {
